requirements.txt
flask
gunicorn
mysql-connector-python
python-dotenv
requests
loguru
beautifulsoup4